+++ drag-and-drop/hitbox/__tests__/unit/get-reorder-destination-index.spec.ts
@@ -1,93 +1,91 @@
-import { reorderWithEdge } from '../../src/reorder-with-edge';
+import { getReorderDestinationIndex } from '../../src/get-reorder-destination-index';
 
 test('moving relative to self should do nothing', () => {
   expect(
-    reorderWithEdge({
-      list: ['A', 'B'],
+    getReorderDestinationIndex({
       startIndex: 0,
       indexOfTarget: 0,
       closestEdgeOfTarget: 'right',
       axis: 'horizontal',
     }),
-  ).toEqual(['A', 'B']);
+  ).toEqual(0);
 
   expect(
-    reorderWithEdge({
-      list: ['A', 'B'],
+    getReorderDestinationIndex({
       startIndex: 0,
       indexOfTarget: 0,
       closestEdgeOfTarget: 'left',
       axis: 'horizontal',
     }),
-  ).toEqual(['A', 'B']);
+  ).toEqual(0);
 
   expect(
-    reorderWithEdge({
-      list: ['A', 'B'],
+    getReorderDestinationIndex({
       startIndex: 1,
       indexOfTarget: 1,
       closestEdgeOfTarget: 'right',
       axis: 'horizontal',
     }),
-  ).toEqual(['A', 'B']);
+  ).toEqual(1);
 
   expect(
-    reorderWithEdge({
-      list: ['A', 'B'],
+    getReorderDestinationIndex({
       startIndex: 1,
       indexOfTarget: 1,
       closestEdgeOfTarget: 'left',
       axis: 'horizontal',
     }),
-  ).toEqual(['A', 'B']);
+  ).toEqual(1);
 });
 
 test('moving forward', () => {
   expect(
-    reorderWithEdge({
-      list: ['A', 'B', 'C'],
+    getReorderDestinationIndex({
+      // list: ['A', 'B', 'C'],
       // move A to left of B
       startIndex: 0,
       indexOfTarget: 1,
       closestEdgeOfTarget: 'left',
       axis: 'horizontal',
     }),
-    // results in no change
-  ).toEqual(['A', 'B', 'C']);
+    // results in no change: ['A', 'B', 'C']
+  ).toEqual(0);
 
   expect(
-    reorderWithEdge({
-      list: ['A', 'B', 'C'],
+    getReorderDestinationIndex({
+      // list: ['A', 'B', 'C'],
       // move A to right of B
       startIndex: 0,
       indexOfTarget: 1,
       closestEdgeOfTarget: 'right',
       axis: 'horizontal',
     }),
-  ).toEqual(['B', 'A', 'C']);
+    // A moved forward ['B', 'A', 'C']
+  ).toEqual(1);
 });
 
 test('moving backwards', () => {
   expect(
-    reorderWithEdge({
-      list: ['A', 'B', 'C'],
+    getReorderDestinationIndex({
+      // list: ['A', 'B', 'C'],
       // move C to right of B
       startIndex: 2,
       indexOfTarget: 1,
       closestEdgeOfTarget: 'right',
       axis: 'horizontal',
     }),
-    // results in no change
-  ).toEqual(['A', 'B', 'C']);
+    // results in no change: ['A', 'B', 'C']
+  ).toEqual(2);
 
   expect(
-    reorderWithEdge({
-      list: ['A', 'B', 'C'],
+    getReorderDestinationIndex({
+      // list: ['A', 'B', 'C'],
       // move C to left of B
       startIndex: 2,
       indexOfTarget: 1,
       closestEdgeOfTarget: 'left',
       axis: 'horizontal',
     }),
-  ).toEqual(['A', 'C', 'B']);
+    // C moved ['A', 'C', 'B']
+  ).toEqual(1);
 });